+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - Remove `INCLUDE_CSS`
 - Remove `INCLUDE_JS`
 - Remove combined.min.css and combined.min.js from static directory
+- Use `[posts.author]` config section instead of `[author]`
 
 ## [6.1.0] - 2025-12-20
 ### Added